Move `get_dimensions_type` to DynamicQuantities interface module
